requirements.txt
baidu_aip==4.16.13
Flask==2.0.0
Flask_Cors==5.0.0
langchain_community==0.3.15
langchain_core==0.3.31
langchain_openai==0.3.1
librosa==0.10.2.post1
numpy==2.2.2
opencv_python==4.10.0.84
pydub==0.25.1
PyMySQL==1.1.0
scipy==1.15.1
tenacity==9.0.0
Werkzeug==2.0.0
